--- src/main/java/com/jkshian/arms/ApplicationController.java
@@ -1,12 +1,0 @@
-package com.jkshian.arms;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class ApplicationController {
-    @GetMapping("/index")
-    public String goHome(){
-        return "index";
-    }
-}
