+++ src/main/java/com/jkshian/arms/dto/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.jkshian.arms.auth;
+package com.jkshian.arms.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,6 +9,9 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationResponse {
-    private String token;
+public class AuthenticationRequest {
+    private String email;
+    String password;
+
+
 }
