+++ src/main/java/com/jkshian/arms/dto/AuthenticationResponse.java
@@ -1,4 +1,4 @@
-package com.jkshian.arms.auth;
+package com.jkshian.arms.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,9 +9,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthenticationRequest {
-    private String email;
-    String password;
-
-
+public class AuthenticationResponse {
+    private String token;
 }
