+++ src/main/java/com/jkshian/arms/service/AuthenticationService.java
@@ -1,9 +1,12 @@
-package com.jkshian.arms.auth;
+package com.jkshian.arms.service;
 
 import com.jkshian.arms.User.Role;
-import com.jkshian.arms.User.User;
-import com.jkshian.arms.User.UserRepository;
+import com.jkshian.arms.entity.User;
+import com.jkshian.arms.repo.UserRepository;
 import com.jkshian.arms.config.JwtService;
+import com.jkshian.arms.dto.AuthenticationRequest;
+import com.jkshian.arms.dto.AuthenticationResponse;
+import com.jkshian.arms.dto.RegisterRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -39,7 +42,7 @@ public class AuthenticationService {
         authenticationManager.authenticate(
                 new UsernamePasswordAuthenticationToken(
                         request.getEmail(),
-                        request.getEmail()
+                        request.getPassword()
                 )
         );
         var user = repository.findByEmail(request.getEmail())
