+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://localhost:3306/Airline-Reservation-System?createDatabaseIfNotExist=true
 spring.datasource.username=root
-spring.datasource.password=
+spring.datasource.password=20010110
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.show-sql=true
